requirements.txt
Flask==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1

Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Flask-Migrate==4.0.7

cloudinary==1.41.0
python-slugify==8.0.4

requests==2.32.3